+++ src/tables/mod.rs
@@ -29,3 +29,4 @@ pub mod name;
 pub mod os2;
 pub mod post;
 pub mod sbix;
+pub mod vvar;